experiments/functions/cold-start/app/requirements-heavy.txt
azure-functions==1.23.0
requests==2.32.3
pandas==2.2.3
numpy==2.2.4
scipy==1.15.2
scikit-learn==1.6.1
matplotlib==3.10.1
joblib==1.4.2
threadpoolctl==3.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
urllib3==2.3.0
charset-normalizer==3.4.1
idna==3.10
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
packaging==24.2
pillow==11.2.1
pyparsing==3.2.3
six==1.17.0
importlib-resources==6.5.2
jinja2==3.1.6
markupsafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
werkzeug==3.1.3
openpyxl==3.1.5
et-xmlfile==2.0.0
